+++ assets/changelog.md
@@ -19,7 +19,7 @@
 * Species accumulation chart in inventory details.
 * Species accumulation chart for selected inventories.
 * Chart comparing the number of species between inventories.
-* Now the user can select the country of species search (Brazil and Uruguay available, more countries to come).
+* Now the user can select the country of species search (Argentina, Brazil, Paraguay and Uruguay available, more countries to come).
 
 ### Improvements